feat:gestion des pages en fonction du status loggedIn + mise en place du logOut

+++ delete-article.php
@@ -1,16 +1,26 @@
 <?php 
 
-    $articles = [];
-    /**
-     * @var ArticleDAO
-     */
-    $articleDAO = require_once './database/models/ArticleDAO.php';
+    require __DIR__.'/database/database.php';
+    $authDAO = require './database/models/AuthDAO.php';
+    $currentUser= $authDAO->isLoggedIn();
 
-    $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-    $idArticle = $_GET['id'] ?? '';
+    if(!$currentUser) {
+        header('Location: /auth-login.php');
+    }else {
+        /**
+         * @var ArticleDAO
+         */
+        $articleDAO = require_once './database/models/ArticleDAO.php';
+        $articles = [];
+
+        $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+        $idArticle = $_GET['id'] ?? '';
 
 
-    if($idArticle) {
-        $articleDAO->deleteOne($idArticle);
+        if($idArticle) {
+            $articleDAO->deleteOne($idArticle);
+        }
+        header('Location: /');
     }
-    header('Location: /');
+
+
